Remove no longer used deprecated request and related utilities
